Split interp_option_to_string/interp_result_to_string static tables into grouped or_else chains and file-split mod_p4_b.rs under the max-lines budget
